Backend validation

+++ server/routes/bills.js
@@ -8,11 +8,20 @@ const connection = mysql.createConnection({
   
 })
 
-const {getEmail} = require('../helper-functions/functions')
+const {getEmail, validateEmail} = require('../helper-functions/functions')
 
 module.exports.addBill = async(req, res) => {
     let {bill_name, bill_source, pay_frequency, next_due, amount, account_id, budget_id} = req.body;
     let email = getEmail(req.headers.authorization).then((email) => {return email;});
+
+    /*if(!bill_name) return res.status(400).json('Bill Name Not Valid');
+    if(!bill_source) return res.status(400).json('Bill Source Not Valid');
+    if(!pay_frequency) return res.status(400).json('Bill Frequency Not Valid');
+    if(!next_due) return res.status(400).json('Bill Date Not Valid');
+    if(!amount) return res.status(400).json('Bill Amount Not Valid');
+    if(!account_id) return res.status(400).json('Account ID Not Valid');
+    if(!budget_id) return res.status(400).json('Budget ID Not Valid');
+    if(!email) return res.status(400).json('Account Email Not Valid');*/
 
     // since the code is set up to take a list, we still have to pass in the single value in list string format
     var idList = budget_id + ", ";
@@ -44,6 +53,7 @@ module.exports.addBill = async(req, res) => {
 module.exports.getBills = async(req, res) => {
     user_id = req.params.user_id;
     let email = getEmail(req.headers.authorization).then((email) => {return email;});
+    if(!email) return res.status(400).json('Account Email Not Valid');
     var sql = `SELECT * FROM Bills INNER JOIN Users ON Users.user_id = Bills.user_id WHERE email = '${await email}'`;
   
     connection.query(sql, function(err, rows)
@@ -84,10 +94,17 @@ module.exports.updateBill = async(req, res) => {
     bill_id = req.params.bill_id;
 
     let {account_id, bill_name, bill_source, amount, next_due, pay_frequency, budget_id} = req.body;
-    console.log(req.body)
     let email = getEmail(req.headers.authorization).then((email) => {return email;});
+    /*if(!bill_name) return res.status(400).json('Bill Name Not Valid');
+    if(!bill_source) return res.status(400).json('Bill Source Not Valid');
+    if(!pay_frequency) return res.status(400).json('Bill Frequency Not Valid');
+    if(!next_due) return res.status(400).json('Bill Date Not Valid');
+    if(!amount) return res.status(400).json('Bill Amount Not Valid');
+    if(!account_id) return res.status(400).json('Account ID Not Valid');
+    if(!budget_id) return res.status(400).json('Budget ID Not Valid');
+    if(!email) return res.status(400).json('Account Email Not Valid');*/
   
-    var sql = `UPDATE Bills SET account_id = '${account_id}', bill_name = '${bill_name}', bill_source = '${bill_source}', amount = '${amount}', next_due = '${next_due}', pay_frequency = '${pay_frequency}' 
+    var sql = `UPDATE Bills SET account_id = '${account_id}', budget_id = '${budget_id}, bill_name = '${bill_name}', bill_source = '${bill_source}', amount = '${amount}', next_due = '${next_due}', pay_frequency = '${pay_frequency}' 
     WHERE bill_id = '${bill_id}' AND user_id=(SELECT user_id FROM Users WHERE email = '${await email}')`;
 
     connection.query(sql, function(err, rows)
@@ -106,7 +123,8 @@ module.exports.updateBill = async(req, res) => {
 module.exports.deleteBill = async(req, res) => {
     bill_id = req.params.bill_id;
     let email = getEmail(req.headers.authorization).then((email) => {return email;});
-
+    /*if(!bill_id) return res.status(400).json('Bill ID Not Valid');
+    if(!email) return res.status(400).json('Account Email Not Valid');*/
     var sql = `DELETE FROM Bills WHERE bill_id = '${bill_id}'
     AND user_id=(SELECT user_id FROM Users WHERE email = '${await email}')`;
